Pre-compile auspice and run the non-dev server by default
This provides a better experience for the `nextstrain view` command.
The dev server has to compile the source each time it is run and uses
hot-reloading in the browser, both of which add overhead and slow down
seeing results.  Those features are unnecessary in the common use case
for `nextstrain view`.

This change relies on a copy of auspice as of 1651d83.

Resolves #2.