Tag each image built from master with a timestamp
This lets us somewhat more easily refer to a specific image version if
we so desire, although the mechanics for picking an image are not great
and timestamps don't carry much meaning about what's in the image.

Note that there is a risk of conflict if two Travis builds are started
in the same second; the last build to finish will "win".  This seems
acceptable, especially for now.  Adding milliseconds to the timestamp
(with ".%3N" after "%S") would reduce the chance of conflict greatly,
but make the tag even more unwieldy.

This is my re-implementation of Trevor's PR #11.  It uses the devel/
scripts I factored out of the Travis config to make builds simpler and
switches to an ISO 8601-compatible timestamp format.