Install augur and auspice from their release branches
We really only want this image to use stable versions of both.  This was
previously mostly happening for auspice because image rebuilds were
triggered only on pushes to the release branch; the master branch was
often equivalent at that time, though it wasn't guaranteed.

If development images need to be built with development versions of
either augur or auspice, the Dockerfile can be amended then to take the
git ref names for each as an image ARG (that defaults to "release" for
both).