Add an entrypoint for AWS Batch jobs
This isn't a true container entrypoint — the standard entrypoint still
execs into this one — but it acts as an extension of the entrypoint for
AWS Batch jobs.  It is responsible for setting up the working directory
from a remote source, running the passed command, and uploading results
back to a remote source.

The exit status of the passed command is passed through, which is useful
because the AWS Batch job metadata includes the job exit status, which
we can use in nextstrain-cli.  This also makes the behaviour of this
entrypoint more like a traditional chained exec, even though that's not
what happens.

Currently only S3 remote sources are supported, but support could be
extended in the future to include HTTP or git.