Update image for Auspice ≥1.35.2
Auspice ≥1.35.2 includes a CLI which should be used to run Auspice going
forward.  The wrapper script is unnecessary and no longer works anyhow
because support for `npm run server` was removed from Auspice.  Using
`npm link` preserves the ability to overlay the image's Auspice version
while still making it globally accessible, mirroring the same strategy
used with Augur.  During development in the Nextstrain ecosystem, it can
be nice to overlay components in the image.

There is a corresponding new-auspice branch for the Nextstrain CLI as
well.