Run basic integration tests on Travis CI
Uses the Nextstrain CLI and the Zika tutorial to kick the tires of the
image.  If these simple things don't work, then the image won't be
published and folks using the CLI won't get a bad image on their next
update.  This, for example, would have caught the Auspice 1.35.2
regression.