Use GitHub Actions for build/pull/tag/push
This replaces the existing scripts with GitHub Actions provided by Docker.

Also has the benefit of using buildx which supports building multi-arch images (to be done later): https://docs.docker.com/desktop/multi-arch/

Sequence of steps inspired by the following guides:

- https://github.com/docker/build-push-action/blob/master/docs/advanced/multi-platform.md
- https://github.com/docker/build-push-action/blob/master/docs/advanced/isolated-builders.md

Not sure if isolated builders are needed here though. It might be more efficient to have one build and push to both base and builder images, if that's even possible?
Or, to cut some time, build the two images in parallel using 2 jobs in this workflow.