ci: Disable pull_request trigger
The PR trigger is unnecessary since all it does is pull/build, which is already done in the checks run by the push trigger.

Also, instead of conditioning on pull_request trigger, conditioning on the repository being the nextstrain repo will ensure the push trigger runs successfully since fork PRs are not expected to have access to nextstrain-bot Docker credentials.