Dockerfile: Compile RAxML from a fork repo to support arm64
The repo contents of RAxML 8.2.12 do not support compiling on arm64. A
few immediate changes are required, so put them in a fork repo.

It's also not likely that these changes will make it back to the
official RAxML repo with a version release, since development efforts
have moved to RAxML-NG¹.

¹ https://groups.google.com/g/raxml/c/1cRSPXcZa1o/m/SakFeA3OAgAJ