copy-images: Use skopeo to "push" images
Instead of pushing images that are available locally (different from a
local registry) to Docker Hub, the proper way to make them available in
the destination is to copy from the source registry directly without
unnecessary pull/push. This is what skopeo can do.

A containerized version of skopeo is used since the version
pre-installed on the GitHub Actions Ubuntu runners is stuck in the past¹
and does not support --multi-arch², which may be used in the future.

Also note that the tag-latest script is replaced by a conditional block
in the new script which copies the tag from one registry to `latest` on
another another.

¹ https://github.com/actions/runner-images/issues/6180#issuecomment-1245858668
² https://github.com/containers/skopeo/releases/tag/v1.6.0