copy-images: Remove inapplicable RepoDigests check
These were kept when moving devel/push to devel/copy-images. However,
checking RepoDigests is only possible if the images are retained in the
local Docker host (different from a local registry). That is not the
case anymore with buildx build --push.

An equivalent error should arise from copy-image() if the image is not
first pushed to a local registry.