Partially revert «Run as a non-root "nextstrain" user by default»
This partially reverts commit b0612f744c9c68d0ec4e56e181e90a1cc345f50d.

The intent of setting a non-root USER was to cover ad-hoc uses of the
image with the best practice of not running as root inside the
container.  However, this got in the way of downstream images (e.g.
nextstrain/ncov-ingest) which inherit from nextstrain/base to add their
own layers on top, as suddenly they changed to running Dockerfile
commands (e.g. "RUN apt install …") as "nextstrain" instead of "root".¹
Since the use by Nextstrain CLI is unaffected either way, just revert
the USER switch and fix the regression for downstream images.

Leaves the user (and group) present so it can be optionally used, though
I'm not sure it will be in practice.  Maybe our entrypoint can `su
nextstrain` in the future?

¹ https://bedfordlab.slack.com/archives/C01LCTT7JNN/p1663797900432129