Install Node.js version 14
Without this change, Node.js v10 is installed. That version reached end
of life on 2021-04-30 [1]. Node.js v14 is the oldest version that is
currently maintained.

Use this in (1) the builder stage to install/build Auspice, and (2) the
final stage to make Auspice usable with the same Node.js version.

Note that apt-get install npm is not necessary since the nodejs package
contains both the node and npm binaries.

[1]: https://github.com/nodejs/Release#end-of-life-releases