Move gofasta, minimap2 to "download" section
These binaries are currently placed in the middle of pip installs. While
they are relevant to the Pangolin packages, this deviates from existing
organization of the Dockerfile which follows:

1. [builder] source builds
2. [builder] package manager installs
3. [final] direct downloads

Add notes that these are for ncov/Pangolin. Also download them in a way
that is consistent with other direct downloads.

Since these commands are no longer bound by the CACHE_DATE declaration
in the builder stage, they can take advantage of caching.