Pin pysam==0.19.1
pysam wasn't pinned in c8466d2f0524d00e2e930765dda58a7e8363e37e
presumably because, when pinned with a git URL, it tries to build from
source and fails due to not having cython installed. However, pinning to
the version allows the pre-built distribution to be used directly.