Prepare for cross-compilation
Split the builder stage into two stages, and install xx¹ which provides
tools to support cross-compilation.

Note that the first stage doesn't need Python so debian:11-slim is used
as the base image.

There is additional work to be done on individual commands to have
proper cross-compilation, but those fixes will come in subsequent
commits to keep them atomic.

¹ https://github.com/tonistiigi/xx