Install g++ to build iltorb
Without this, there is a build error:

      CXX(target) Release/obj.target/iltorb/src/common/allocator.o
    make: g++: Command not found

iltorb's Makefile is not in the official GitHub repository, but can be
inspected from an npm install. It looks to be generated by and old
version of gyp¹, and is near identical to this example².

The line `CXX.host ?= g++` is the source of the error. This means it
should be right to install native g++ as it should be used for "host"
purposes only, not to build for a different target architecture.

¹ https://chromium.googlesource.com/external/gyp/+/54d2f6fe6d8a7b9d9786bd1f8540df6b4f46b83f%5E%21/pylib/gyp/generator/make.py
² https://github.com/chentsulin/node-addon-examples/blob/master/pass-parameters/build/Makefile