build: Use intermediate registry as a cache source
Currently, only images available locally or on the destination
registry (Docker Hub) are used for caching. This was fine before the
usage of an intermediate directory. Now that built images are
no longer available locally, there are two new issues:

1. The final stage does not use the cached builder stage since it is no
   longer available from cache sources.
2. Local development is slower since the built image layers are not
   available from cache sources.

Adding the intermediate registry as a cache source fixes these.