Downgrade Node.js to v14
Node.js v16 is very slow when running in an arm64 Docker build on an amd64
build platform.

Node.js v14 runs in the build without emulation. It uses emulation in
the Docker container during run time, but the slowness there is not
nearly as significant as the build time increase due to emulation.

This should be reverted if/when cross-compilation is used, i.e. Node.js
v16 can run natively on the build platform while installing packages for
a different target platform.

This reverts commit 1afd07dae89be39b145ab465d42f5fd8b12cedd8.