Add the start of a test suite of image properties and behaviours
These are more like unit tests of basic, foundational properties of the
image, and they complement the broader integration tests we already do
in CI (e.g. running pathogen builds).

Starting new containers lots of times makes the tests pretty darn slow,
but I think it'll be fine for now and for a long while.  We can address
it in the future, when we start getting too aggravated by it.  Options
to do so include launching a single container per test file and using
`docker exec` to run commands in it (although this changes semantics too
so isn't a drop-in replacement) and/or running each test file in
parallel to the others.