build: Default the Docker platform based on the Docker server's arch
The goal is to make builds "native by default" in development, for
faster iteration times.

Using the Docker server's arch is more correct than the current
machine's arch (i.e. `uname -m`) since we use Buildx to build within a
container and the container will run on the server, which may be
different than the current machine.