Omit dev deps when installing Auspice
This avoids disk-heavy deps like Puppeteer, which bundle a whole
Chromium install.  Based on some rough comparisons, I expected this to
shave about 600MB from the uncompressed image size and a local test
build bore that out.

The node_modules/ tree is infamously large and bloated.  The usual
culprits are non-source files that are commonly included in package
distributions but not needed at run time.  So there is surely more we
could shave off here, but this is a huge easy start.