start-localhost-registry: Start container if it already exists
Previously an error occurred if the container existed because container
creation tried to use a name already in use.  It's safe to "start" a
container that's already started, so this makes the program idempotent.

This also handles the case of the container existing but being stopped,
which previously wasn't handled.