Remove unnecessary and unwarranted `npm update` during Auspice install
This was introduced incidentally, and seemingly unintentionally¹, in
"Replace Alpine-based image with a Debian-based image" (ce07dad), but no
one noticed.

`npm update` is unwarranted because it's intended for maintainers of a
package, not downstream user installs, which is the role we have here.
Its effect was to bump the minimum versions in Auspice's package.json to
the latest available (while still respecting SemVer constraints) and
then install all of Auspice's deps into node_modules/.²  That's
unnecessary because we then run `npm install`, which unlike `npm
update`, also runs pre/post-installation steps Auspice includes.

¹ <https://github.com/nextstrain/docker-base/pull/21#discussion_r1239131111>

² Notably, `npm update` was added when the image had npm v5, which
  updates both package.json and package-lock.json.  We currently have npm
  v6, which does the same.  Subsequent versions, e.g. npm v8, stopped
  updating package.json and only update package-lock.json.