Update jaxlib for evofr on linux/arm64
The latest release of evofr included updates to its dependencies that
updated the minimum version for jaxlib.¹ I'm using the latest release
of jaxlib available in the user forked repo since there's still no
official pre-built binaries for linux/arm64.²

I did not realize this change was needed until I saw the
"Validate Platform" job fail in the CI run³ that I triggered to update
evofr to v0.1.20. Added a note for our future selves to check the
jaxlib installation if we run into different versions of evofr on the
different platforms.

¹ https://github.com/blab/evofr/commit/46f744b605c258d6f3d3b7b3ac8bc877f8e4ba22
² https://github.com/google/jax/issues/7097
³ https://github.com/nextstrain/docker-base/actions/runs/5480222698