Pin jax version for evofr on linux/arm64
Previously, we were only pinning jaxlib because the there's no official
pre-built binaries for linux/arm64. However, jax has jaxlib minimum
version requirements that are _not_ checked by package version
constraints in pip.¹ Instead, the jaxlib version is checked at runtime.
This led to a mismatch of versions when only pinning jaxlib.²

Instead of always checking the minimum version requirements for jaxlib³
it's easier to just install the latest version of jaxlib available in
the user forked repo and pin jax to the matching version.

¹ https://jax.readthedocs.io/en/latest/jep/9419-jax-versioning.html#how-are-jax-and-jaxlib-versioned
² https://github.com/nextstrain/forecasts-ncov/pull/50#issuecomment-1629841748
³ https://github.com/google/jax/blob/main/jax/version.py