Isolate AWS CLI (v1) into its own Python venv
Otherwise, the awscli → botocore dependency is broken by the subsequent
installation of the nextstrain-cli → botocore dependency, which results
in --no-sign-request not working as it should.

This approach seemed like the option of least change.  Alternatives to
and/or additional improvements to this include a) installing Nextstrain
CLI in its own venv instead/as well, b) upgrading to AWS CLI v2 (which
is always isolated) or c) updating the nextstrain-cli → botocore
dependency to not conflict.

Resolves: <https://github.com/nextstrain/docker-base/issues/213>
Related-to: <https://github.com/nextstrain/docker-base/pull/214>