entrypoint-aws-batch: Upload .snakemake/metadata/ too
Snakemake stores state information per input/output here and uses it to
determine if it needs to re-run rules or not.  It seems akin to the file
mtimes which we already take care to preserve on upload/download.
Additionally, the metadata recorded is used in Snakemake's report
generation and is generally useful for looking at workflow statistics.

Continue to not upload all of .snakemake/ en masse because it can
potentially contain files that interfere with local usage and/or are
large and unnecessary.

Resolves: <https://github.com/nextstrain/cli/issues/373>
Related-to: <https://github.com/nextstrain/cli/pull/374>