CI: Update test-pathogen-repo-ci
Add comments to clearly show the job using pathogen-repo-ci@v0 is for
pathogen repos that do not conform to the standard pathogen repo
structure. Since these are tied to a old tag of the `pathogen-repo-ci`,
the job should not be expected to be updated.

Any new pathogen repos should be added to the `test-pathogen-repo-ci`
job which uses the latest version of `pathogen-repo-ci`.