Upgrade IQ-TREE to 2.3.5 with CMAPLE support
We want the CMAPLE support (first available with 2.3.4) for better and
faster trees.

CMAPLE support is optional and not all release builds of IQ-TREE have it
enabled.  So far, it seems that the "main" releases are made without
CMAPLE support (e.g. 2.3.5) and then parallel releases are made with it
(e.g. 2.3.5.cmaple), but only for x86_64 on Linux and macOS.  Thus,
we're still not using an aarch64/arm64 binary even though the "main"
releases (without CMAPLE support) provide them.  If we keep waiting,
maybe they'll start being provided, or, eventually, we can start
compiling them ourselves.  That's for another time, though.

Resolves: <https://github.com/nextstrain/docker-base/issues/226>