Upgrade base image from Debian Bullseye (oldstable) to Bookworm (stable)
Avail ourselves of newer (though still not _new_ ;) software, in
particular the SQLite CLI.

The use of Bullseye's version number, 11, instead of its name in the
builder-build-platform image tripped me up at first, so I've switched it
over to Bookworm's name instead of number (12).

The cross-compilation of vcftools targeting arm64/aarch64 on amd64
requires pkg-config installed for the *target* platform with xx-apt-get.
Without it, the vcftool's configure script can't find zlib since its
using the *build* platform's pkg-config and zlib is not installed for
the build platform.  I'm not sure why this worked with the build
platform's pkg-config previously but not now.

The libopenblas-base package was renamed libopenblas0.  In Bullseye, the
former was a dependency-only package to allow older systems coming from
Buster (oldoldstable), or instructions based on it, to more gracefully
transition to Bullseye.  The transitional package was removed in
Bookworm.