Run `nextstrain check-setup` as nextstrain:nextstrain instead of root
Otherwise when `augur` is invoked by `nextstrain check-setup`, it's the
first time and so Python creates __pycache__/**.pyc bytecode files for
Augur's modules and they're owned by root:root.  This makes the files
impossible to remove/update as the nextstrain:nextstrain user inside the
container, which is important for overlays.  It's also counter to the
expectation that everything under /nextstrain is owned by
nextstrain:nextstrain.  Switching user in the Dockerfile is more
future-proof than adding another `chmod -R`.