entrypoint-aws-batch: Remove existing /nextstrain/{augur,auspice,fauna}/ for overlays
Makes these overlays work the same on AWS Batch as they do in local
containers (Docker or Singularity): the overlays fully replace the
image's copy.  This addresses a caveat noted the Nextstrain CLI commit
adding AWS Batch overlay support¹ that was further discussed in review.²

ZIP archive directory members always end in "/", so we test only for the
presence of ../{augur,auspice,fauna}/ in the archive (or the presence of
equivalent relative paths adjusted for a non-default workdir).  This
works consistently for workdir archives with overlays produced by
Nextstrain CLI, but it isn't foolproof for archives produced other ways.
For example, it is possible for a ../augur/x/y/z path to be present in
the archive without ../augur/ also being present.  In that case, nothing
will be deleted.  This could be seen as a bug or a feature.

¹ <https://github.com/nextstrain/cli/commit/99168ac05502eebbef806e11f891b676b7e6f8cd>
² <https://github.com/nextstrain/cli/pull/419#discussion_r1982432118>