Replace fastparquet with pyarrow
We need support for reading/writing parquet files to prepare submissions
to the SARS-CoV-2 variant hub [1]. The pyarrow library is one of two
supported by pandas [2] along with fastparquet. The pyarrow library provides
a more comprehensive set of tools for the Arrow spec [3], while
fastparquet is defined to provide a minimal library for the parquet
format. We need to switch to the larger pyarrow library here, because it
supports the parquet DATE data type that we need for our SARS-CoV-2
nowcast submissions.

[1] https://github.com/nextstrain/forecasts-ncov/issues/132
[2] https://pandas.pydata.org/pandas-docs/stable/user_guide/io.html#io-parquet
[3] https://arrow.apache.org/docs/cpp/user_guide.html