Add support for reading/writing parquet files
We need support for reading/writing parquet files to prepare submissions
to the SARS-CoV-2 variant hub [1]. The pyarrow library is one of two
supported by pandas [2] along with fastparquet. The pyarrow library provides
a more comprehensive set of tools for the Arrow spec [3], while
fastparquet is defined to provide a minimal library for the parquet
format. I've opted for the larger pyarrow library here, since it will
eventually be a required dependency for pandas [4].

[1] https://github.com/nextstrain/forecasts-ncov/issues/132
[2] https://pandas.pydata.org/pandas-docs/stable/user_guide/io.html#io-parquet
[3] https://arrow.apache.org/docs/cpp/user_guide.html
[4] https://pandas.pydata.org/pdeps/0010-required-pyarrow-dependency.html